Migrate zip to cp --zip
