+++ src/lightning/app/cli/commands/cp.py
@@ -19,11 +19,18 @@ from functools import partial
 from multiprocessing.pool import ApplyResult
 from pathlib import Path
 from typing import Any, Optional, Tuple, Union
+from textwrap import dedent
 
 import click
 import requests
 import urllib3
-from lightning_cloud.openapi import Externalv1LightningappInstance, ProjectIdStorageBody, V1CloudSpace
+from lightning_cloud.openapi import (
+    Externalv1Cluster,
+    Externalv1LightningappInstance,
+    ProjectIdStorageBody,
+    V1CloudSpace,
+    V1GetClusterResponse,
+)
 from rich.live import Live
 from rich.progress import BarColumn, DownloadColumn, Progress, TaskID, TextColumn
 from rich.spinner import Spinner
@@ -33,6 +40,7 @@ from lightning.app.cli.commands.ls import _collect_artifacts, _get_prefix
 from lightning.app.cli.commands.pwd import _pwd
 from lightning.app.source_code import FileUploader
 from lightning.app.utilities.app_helpers import Logger
+from lightning.app.utilities.auth import _AuthTokenGetter
 from lightning.app.utilities.cli_helpers import _error_and_exit
 from lightning.app.utilities.network import LightningClient
 
@@ -43,7 +51,8 @@ logger = Logger(__name__)
 @click.argument("dst_path", required=True)
 @click.option("-r", required=False, hidden=True)
 @click.option("--recursive", required=False, hidden=True)
-def cp(src_path: str, dst_path: str, r: bool = False, recursive: bool = False) -> None:
+@click.option("--zip", required=False)
+def cp(src_path: str, dst_path: str, r: bool = False, recursive: bool = False, zip: bool = False) -> None:
     """Copy files between your local filesystem and the Lightning Cloud filesystem."""
 
     if sys.platform == "win32":
@@ -54,9 +63,6 @@ def cp(src_path: str, dst_path: str, r: bool = False, recursive: bool = False) -
 
         pwd = _pwd()
 
-        if pwd == "/" or len(pwd.split("/")) == 1:
-            return _error_and_exit("Uploading files at the project level isn't allowed yet.")
-
         client = LightningClient(retry=False)
 
         src_path, src_remote = _sanitize_path(src_path, pwd)
@@ -66,8 +72,14 @@ def cp(src_path: str, dst_path: str, r: bool = False, recursive: bool = False) -
             return _error_and_exit("Moving files remotely isn't supported yet. Please, open a Github issue.")
 
         if not src_remote and dst_remote:
+            if dst_path == "/" or len(dst_path.split("/")) == 1:
+                return _error_and_exit("Uploading files at the project level isn't allowed yet.")
+            if zip:
+                return _error_and_exit("Zipping uploads isn't supported yet. Please, open a Github issue.")
             _upload_files(live, client, src_path, dst_path, pwd)
         elif src_remote and not dst_remote:
+            if zip:
+                return _zip_files(live, src_path, dst_path)
             _download_files(live, client, src_path, dst_path, pwd)
         else:
             return _error_and_exit("Moving files locally isn't supported yet. Please, open a Github issue.")
@@ -160,6 +172,45 @@ def _upload(source_file: str, presigned_url: ApplyResult, progress: Progress, ta
     file_uploader.progress = progress
     file_uploader.task_id = task_id
     file_uploader.upload()
+
+
+def _zip_files(live: Live, remote_src: str, local_dst: str) -> None:
+    live.stop()
+
+    if len(remote_src.split("/")) < 3:
+        return _error_and_exit(
+            dedent(
+                f"""
+                The source path must be at least two levels deep (e.g. r:/my-project/my-lit-resource).
+
+                The path provided was: r:{remote_src}
+                """
+            )
+        )
+
+    if os.path.isdir(local_dst):
+        local_dst = os.path.join(local_dst, os.path.basename(remote_src) + ".zip")
+
+    project_id, lit_resource = _get_project_id_and_resource(remote_src)
+
+    # /my-project/my-lit-resource/artfact-path -> cloudspace/my-lit-resource-id/artifact-path
+    artifact = "/".join(remote_src.split("/")[3:])
+    prefix = _get_prefix(artifact, lit_resource)
+
+    token = _AuthTokenGetter(LightningClient().api_client)._get_api_token()
+    endpoint = f"/v1/projects/{project_id}/artifacts/download?prefix={prefix}&token={token}"
+
+    cluster = _cluster_from_lit_resource(lit_resource)
+    url = _storage_host(cluster) + endpoint
+
+    progress = _get_progress_bar(transient=True)
+    progress.start()
+    task_id = progress.add_task("download zip", total=None)
+
+    _download_file(local_dst, url, progress, task_id)
+    progress.stop()
+
+    click.echo(f"Downloaded to {local_dst}")
 
 
 def _download_files(live, client, remote_src: str, local_dst: str, pwd: str):
@@ -284,3 +335,23 @@ def _get_progress_bar(**kwargs: Any) -> Progress:
         DownloadColumn(),
         **kwargs,
     )
+
+
+def _storage_host(cluster: Union[V1GetClusterResponse, Externalv1Cluster]) -> str:
+    dev_host = os.environ.get("LIGHTNING_STORAGE_HOST")
+    if dev_host:
+        return dev_host
+    return f"https://storage.{cluster.spec.driver.kubernetes.root_domain_name}"
+
+
+def _cluster_from_lit_resource(
+    lit_resource: Union[Externalv1LightningappInstance, V1CloudSpace]
+) -> Union[V1GetClusterResponse, Externalv1Cluster]:
+    client = LightningClient()
+    if isinstance(lit_resource, Externalv1LightningappInstance):
+        return client.cluster_service_get_cluster(lit_resource.spec.cluster_id)
+
+    clusters = client.cluster_service_list_clusters()
+    for cluster in clusters.clusters:
+        if cluster.id == clusters.default_cluster:
+            return cluster
